Initial refactor to support single protocol for hub scenes and external scenes
