+++ custom_components/stateful_scenes/const.py
@@ -1,5 +1,9 @@
 """Constants for the State Scene integration."""
 
+from typing import Any, Protocol
+
+from homeassistant.helpers.entity_registry import ReadOnlyDict
+
 DOMAIN = "stateful_scenes"
 
 # Hub configuration
@@ -66,3 +70,19 @@ class StatefulScenesYamlNotFound(Exception):
 
 class StatefulScenesYamlInvalid(Exception):
     """Raised when specified yaml is invalid."""
+
+
+class SceneStateAttributes(ReadOnlyDict[str, Any]):
+    """Protocol Attributes for HA/external scenes or hub state."""
+
+    friendly_name: str
+    icon: str | None
+    area_id: str | None
+    entity_id: list[str]
+
+
+class SceneStateProtocol(Protocol):
+    """Protocol for HA and/or integration State."""
+
+    @property
+    def attributes(self) -> SceneStateAttributes: ...  # noqa: D102
